Update TestNetscapePlugIn to build 64-bit using the Cocoa event model.
It currently does not attempt to print events which means that plugins/mouse-events.html
will fail when run 64-bit.  All other tests that use this plugin pass.

Reviewed by Anders Carlsson.

* DumpRenderTree/TestNetscapePlugIn.subproj/PluginObject.cpp:
(testGetIntIdentifier):
* DumpRenderTree/TestNetscapePlugIn.subproj/main.cpp:
(NPP_New):
(NPP_HandleEvent):


Canonical link: https://commits.webkit.org/26209@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@33019 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebKitTools/DumpRenderTree/TestNetscapePlugIn.subproj/PluginObject.cpp
@@ -401,7 +401,7 @@ static bool testGetIntIdentifier(PluginObject*, const NPVariant* args, uint32_t 
         return false;
 
     NPIdentifier identifier = browser->getintidentifier((int)NPVARIANT_TO_DOUBLE(args[0]));
-    INT32_TO_NPVARIANT((int32)identifier, *result);
+    INT32_TO_NPVARIANT((int32)(long long)identifier, *result);
     return true;
 }
